backend/requirements.txt
python-nmap
gvm-tools
django
netifaces
xmltodict
django-cors-headers
